This is the pipeline for SLE analysis
I am creating this pipeline to take PE50 reads generated from Cut&Run_seq and bulk RNA-seq. The pipeline is ran in the slurm job scheduler and requires a txt file that contains 4 columns (see ELF1_Mock_CutRun_kd_ctr.txt). The first two columns should have the forward and reverse reads for the knockdown replicate and the second two columns contain the forward and reverse reads for the control replicates. We will have multiple text files each containing this format for each condition and transcription factor. I will add a .txt file to show as an example. The .txt file will be in the same directory as all the fastq.gz files.